+++ changelog.md
@@ -5,6 +5,8 @@
 1. Respect launch params when setting host, port and workers
 2. Switched to use https protocol for steam api requests
 3. Added `last_seen_before` param to **GET** `/member/list`
+4. Moved **GET** `/divisions` to **GET** `/divisions/statistics`
+5. Added **GET** `/divisions/{divisionid}/activity` for per-user statistics
 
 ## v2.10.0